install rhosp-director-images-ipa from url
Install rhosp-director-images-ipa package directly from the repo
url to bypass need of dependencies.
This will prevent to tag more packages, since using downloadonly
option of yum/dnf still requires to have the dependencies
available.
At the moment the package is unfortunately tagged in 2 repos,
so we need to grep for the correct one.